Adjusted headings style; Renamed component
